Remove dependency on `alpine` image
If the controller is running in a airgapped environment
it will not have access to the internet to download the
`alpine` image. This change uses the builder image to run
the commands of the prepare step.

This approach as a caveat. If the builder image does not have
`chmod` installed it will fail.

Signed-off-by: Joao Pereira <jdealmeidapereira@pivotal.io>