Correct ClusterBuilder metadata retrieval keychain
- We need to create a Keychain and start to use it as a multi keychain
and read from the provider that we implemented. We might be able to get away with it as it is a corner case that shows up when the registry your images are in is the same GCP account as the cluster.
- Also ggcr is forking the providers and we should update our version of
ggcr when they finish this work so we do not have to depend on k8s.net/kubernetes
package
- Remove function NewRemoteWithDefaultKeychain that is no longer needed

Signed-off-by: Ashwin Venkatesh <avenkatesh@pivotal.io>