Update id for yarn
My ccb wasn't building successfully due to yarn ID not being available in the store.  I'd imagine at some point we will have to swap out the IDs for paketo at some point, so maybe just do that en-masse.  If this PR is accepted I'd be happy to make the change in other samples.