Add CNB Service Bindings to Image and Build resources
The CNB Service Bindings spec provides a structured way to expose
service metadata and credentials into a build. For example, the build
can use the binding information to install a database driver, configure
an APM agent for the built image, or discover any other capability
outside of the built container.

Each binding reference has a name, a reference to a k8s ConfigMap
holding service metadata, and optionally a k8s Secret holding service
credentials. Builds generally should not have access to the credentials
so that the values are not backed into the built image, but instead
bound fresh at runtime.

For images, the bindings are located at `.spec.build.bindings`:

```
kind: Image
...
spec:
  ...
  build:
    bindings:
    - name: my-binding
      metadataRef:
        name: provider-binding-metadata
      secretRef:
        name: provider-binding-secret
```

For builds, the bindings are located at `.spec.bindings`:

```
kind: Build
...
spec:
  ...
  bindings:
  - name: my-binding
    metadataRef:
      name: provider-binding-metadata
    secretRef:
      name: provider-binding-secret
```

In the build the bindings appear as a directory tree under the path
specified by the `$CNB_BINDINGS` env var. Each binding is a directory,
which in turn has a `metadata` directory with files and if credentials
are exposed a `secret` directory with more files. The environment
variable and volume mounts are available during the detect and build
steps. Currently, the value of `$CNB_BINDINGS` is `/var/bindings` but
should not be presumed to be stable or consistent with the value at
runtime.

There is no guarantee the same services, or types of services, will be
bound at runtime.

Refs #369