Don't write build cache to emptydir
Previously, when no cache-volume was specified the caching-data got written to an EmptyDir.
EmptyDirs are thrown away once the Pod is finished.
For some buildpacks/apps (e.g. gradle when building spring-music) generating the cache takes a considerable amount of time. If we know that we will throw it away anyway we can skip that.

Therefore we:
- Don't pass the `--cache-dir` arg to the lifecycle steps
- Don't add the cache-dir to the `VolumeMounts`
- Don't create an EmptyDir for the cache

Co-authored-by: Benjamin Haegenlaeuer <benjamin.haegenlaeuer@sap.com>
Co-authored-by: Philipp Stehle <philipp.stehle@sap.com>