Only Filter Controller by Group/Version
This prevents an issue where "owned" resources created with an earlier kpack api version were not triggering reconciles of their owning resource.

An example of this was with the Image resource which relies on re-reconciling when a source resolver is updated. However, if the source resolver was created before a kpack controller update it would not rereconcile the image on updates.