parse short scp-like syntax into proper ssh urls
currently we're lelying on an underlying library (libssh2) to be able to
handle the shorter `[user@]host:path` syntax, but we should be good
citizens and convert it ourselves.

one key thing to note is that we don't convert any syntax that omits the
user (i.e. `host:path`). while git allows this and will use the current
user (i.e. cnb), this is a behaviour we don't want people to rely on.

Signed-off-by: Bohan Chen <bohanc@vmware.com>