add explicit known_hosts support for git ssh
when we switched from libgit2 to go-git, the ssh source was broken
because libgit2 implicitly trusted unknown hosts whereas go-git required
it to be explicitly done in code.

this commit fixes that regression by introducing a flag on the
controller to explicitly trust all unknown hosts. at the same time, we
also introduce the ability to specify the trusted hosts via the git ssh
secret.

Signed-off-by: Bohan Chen <bohanc@vmware.com>