add ability to extract metadata from images
The source metadata is read in from the project-metadata label on the
built image. This is done for 2 reasons:

1. It resolves floating pointers to concrete shas, this can be the case
   with a git source pointing to a branch, or a registry source pointing
   to a tag. We also can't figure out a blob's checksum without actually
   pulling it
2. It guarantees that this is the version pulled down by the `prepare`
   step and avoid weird TOCTOU edge cases where the underlying bits
   changed between the creation of the Build and the execution of the
   container

By exporting the Project, Source, Metadata, and Version structs, I hope
to create a code relation between the different packages, so that if we
every change one in the future, the compiler will yell at us instead of
us having to remember the relationship.

Signed-off-by: Bohan Chen <bohanc@vmware.com>