Update fetch.go to include a Depth specifier (#1732)
* Update fetch.go to include a Depth specifier

For repositories with large histories, since there doesn't appear to be caching of the git repository, we shouldn't pull all history as it's not needed for the purpose of building.  This one line change significantly improves build performance when irrelevant git history is large. 

Signed-off-by: ollie-kane-CB <136009654+ollie-kane-CB@users.noreply.github.com>

* Update fetch.go to include a force flag.

Based on https://github.com/rancher/rancher/issues/42659, it appears that it may be necessary to include a force flag

Signed-off-by: ollie-kane-CB <136009654+ollie-kane-CB@users.noreply.github.com>

* Update fetch.go

Signed-off-by: ollie-kane-CB <136009654+ollie-kane-CB@users.noreply.github.com>

* Update fetch_test.go

Signed-off-by: ollie-kane-CB <136009654+ollie-kane-CB@users.noreply.github.com>

---------

Signed-off-by: ollie-kane-CB <136009654+ollie-kane-CB@users.noreply.github.com>