Do not reconcile objects that are being deleted
If the kpack image (or its builds) has a finalizer being added by an
external component, then when the client request the object to be
deleted, the image/build reconcilers keep reconciling the image/build
causing new build pods to be created as the image is being deleted. Even
though eventually the image gets deleted, it takes significant amount of
time.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>