Do not attempt to download lifecycle binary for Windows
As of lifecycle 0.20.6, Windows support is removed,
and no lifecycle binary is available for Windows on GitHub.

Without this change, we'd be unable to upgrade the lifecycle version in kpack
beyond 0.20.5.

This change avoids attempting to download the Windows binary,
but still produces an empty Windows layer in the lifecycle image,
to minimize the surface area of changes.

This is a first step toward deprecating Windows support in kpack
(see https://github.com/buildpacks-community/kpack/discussions/1366),
which requires further changes throughout the code base
to remove Windows-specific logic entirely.

Signed-off-by: Natalie Arellano <narellano@vmware.com>